feat: refactor into modules
Moves the various stages into their own modules and simplifies the main
entrypoint because of that.

Also adds a `bin` to the `package.json`.
