+++ src/bin.ts
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+import {run} from './main.js';
+
+run();
